% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAssessmentKey.R
\name{findAssessmentKey}
\alias{findAssessmentKey}
\title{Find a assessmentKey}
\usage{
findAssessmentKey(stock, year = 0, regex = TRUE, full = FALSE)
}
\arguments{
\item{stock}{a stock name, e.g. cod-347d, or cod to find all cod stocks, or
NULL to process all stocks.}

\item{year}{the assessment year, e.g. 2015, or 0 to process all years.}

\item{regex}{whether to match the stock name as a regular expression.}

\item{full}{whether to return a data frame with all stock list columns.}
}
\value{
A vector of assessment keys (default) or a data frame if full is TRUE.
}
\description{
Find a lookup key corresponding to a stock in a given assessment year.
}
\examples{
findAssessmentKey("cod-347d", 2015, full = TRUE)

}
\seealso{
\code{\link{getListStocks}} gets a list of stocks.

\code{\link{icesSAG-package}} gives an overview of the package.
}
\author{
Arni Magnusson and Colin Millar.
}
