% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_image.R
\name{get_custom_plot}
\alias{get_custom_plot}
\title{Get a Custom Graph of Stock Assessment Output}
\usage{
get_custom_plot(assessmentKey, type = c(15, 16, 17, 18), width = 800, ...)
}
\arguments{
\item{assessmentKey}{the unique identifier of the stock assessment}

\item{type}{The type of plot: values can be 15, 16, 17, or 18.}

\item{width}{width of the image in pixels}

\item{...}{to allow scope for back compatibility}
}
\value{
An array representing a bitmap.
}
\description{
Get a custom graph of stock assessment output.
}
\examples{
\dontrun{
assessmentKey <- findAssessmentKey("ane.27.9a", year = 2022)
landings_img <- get_custom_plot(assessmentKey, "15")

plot(landings_img)

landings_plots <- get_image(assessmentKey, "landings")
plot(landings_plots)
}

}
\seealso{
\code{\link{getListStocks}} gets a list of stocks.

\code{\link{getFishStockReferencePoints}} gets biological reference points.

\code{\link{icesSAG-package}} gives an overview of the package.
}
