% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_image.R
\name{get_image}
\alias{get_image}
\title{Get a Graph of Stock Assessment Output}
\usage{
get_image(
  assessmentKey,
  type = c("landings", "recruitment", "ssb", "mortality", "historical_mortality",
    "historical_ssb", "historical_recruitment"),
  width = 800,
  ...
)
}
\arguments{
\item{assessmentKey}{the unique identifier of the stock assessment}

\item{type}{The type of plot: values can be landings, recruitment
ssb, mortality, historical_mortality, historical_ssb,
historical_recruitment.}

\item{width}{width of the image in pixels}

\item{...}{to allow scope for back compatibility}
}
\value{
An array representing a bitmap.
}
\description{
Get a graph of stock assessment output, e.g., historical stock size,
recruitment, and fishing pressure.
}
\examples{
\dontrun{
assessmentKey <- findAssessmentKey("cod.27.21", year = 2022:2023)
landings_img <- get_image(assessmentKey[1], "landings")

plot(landings_img)

landings_plots <- get_image(assessmentKey, "landings")
plot(landings_plots)
}

}
\seealso{
\code{\link{getListStocks}} gets a list of stocks.

\code{\link{getFishStockReferencePoints}} gets biological reference points.

\code{\link{icesSAG-package}} gives an overview of the package.
}
