% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLatestStockAdviceList.R
\name{getLatestStockAdviceList}
\alias{getLatestStockAdviceList}
\title{Get List of Most Recent Advice}
\usage{
getLatestStockAdviceList(...)
}
\arguments{
\item{...}{arguments passed to \code{\link{sag_get}}.}
}
\value{
A data frame.
}
\description{
Get a list of the most recent advice for all fish stocks.
}
\examples{
\dontrun{
stocks <- getLatestStockAdviceList()
}
}
\seealso{
\code{\link{getSummaryTable}} gets a summary table of historical stock size.

\code{\link{getFishStockReferencePoints}} gets biological reference points.

\code{\link{icesSAG-package}} gives an overview of the package.
}
\author{
Colin Millar, Scott Large, and Arni Magnusson.
}
