% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSAGGraph.R
\name{getSAGGraph}
\alias{getSAGGraph}
\title{Get a Summary Table of Historical Stock Size}
\usage{
getSAGGraph(key, type)
}
\arguments{
\item{key}{the unique identifier of the stock assessment}

\item{type}{the type of plot requested, e.g. "Landings", "Recruitment", ...}
}
\value{
An array representing a png.
}
\description{
Get a summary table of historical stock size, recruitment, and fishing pressure.
}
\examples{
stocklist <- getListStocks(2016)
id <- grep("cod-2224", stocklist$FishStockName)
stocklist[id,]
key <- stocklist$key[id]
rec_img <- getSAGGraph(key, type = "Recruitment")
plot.new()
plot(rec_img)

}
\author{
Colin Millar and Scott Large.
}
\seealso{
\code{\link{getListStocks}} gets a list of stocks.

\code{\link{getFishStockReferencePoints}} gets biological reference points.

\code{\link{icesSAG-package}} gives an overview of the package.
}

