% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAGtokens.R
\name{getTokenExpiration}
\alias{getTokenExpiration}
\title{Get the expiration time of a Standard Graphs token.}
\usage{
getTokenExpiration()
}
\value{
An integer.
}
\description{
Get the number of days remaining for a standard graphs
personal access token.
}
\examples{
\dontrun{
getTokenExpiration()
}

# tell icesSAG to use the SG token
options(icesSAG.use_token = TRUE)

}
\author{
Colin Millar.
}
