% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesSAG-package.R
\docType{package}
\name{icesSAG-package}
\alias{icesSAG}
\alias{icesSAG-package}
\title{Stock Assessment Graphs (SAG) database webservices}
\description{
Functions that access the webservices of the ICES (International
Council for the Exploration of the Sea) Stock Assessment Graphs (SAG) database.
}
\author{
Colin Millar, Scott Large, and Arni Magnusson.
}
\references{
Overview of the ICES advisory process:
\url{http://ices.dk/community/advisory-process/Pages/default.aspx}.

ICES list of software applications:
\url{http://ices.dk/marine-data/tools/Pages/Software.aspx}.

ICES SAG Webservices:
\url{http://ices.dk/marine-data/tools/Pages/stock-assessment-graphs.aspx}.
}

