% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatestStockAdviceList.R
\name{LatestStockAdviceList}
\alias{LatestStockAdviceList}
\title{Get a List of the Latest Fish Stock Assessments}
\usage{
LatestStockAdviceList(...)
}
\arguments{
\item{...}{arguments passed to \code{\link{ices_get}}.}
}
\value{
A data frame.
}
\description{
Get a list of the latest fish stock assessments.
Only published stocks are returned.
}
\examples{
\dontrun{
stock_list <- LatestStockAdviceList()
}
}
\seealso{
\code{\link{StockList}} Get a list of fish stock assessments.

\code{\link{getFishStockReferencePoints}} gets biological reference points.

\code{\link{icesSAG-package}} gives an overview of the package.
}
