% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSAG.R
\name{writeSAG}
\alias{writeSAG}
\alias{readSAGxml}
\title{Create and read the SAG CSV data transfer file}
\usage{
writeSAG(info, fishdata, file = NULL)

readSAGxml(file)
}
\arguments{
\item{info}{a list of stock information}

\item{fishdata}{a data frame of fish data}

\item{file}{a csv file name}
}
\value{
a string containing the csv file.
}
\description{
Convert between R data (a list and a data.frame) and the CSV format required for uploading
data to the SAG database.
}
\examples{

\dontrun{
info <- stockInfo(StockCode = "cod.27.347d",
                  AssessmentYear = 2017,
                  StockCategory = 1,
                  ModelType = "A",
                  ModelName = "SCA",
                  ContactPerson = "itsme@fisheries.com")
fishdata <- stockFishdata(Year = 1990:2017, Catches = 100)
csvfile <- writeSAG(info, fishdata)
}

}
\seealso{
\code{\link{stockInfo}} creates a list of stock information.

\code{\link{stockFishdata}} creates a data frame of fish stock summary data.
}
