% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sag_api.R
\name{sag_api}
\alias{sag_api}
\title{Build a SAG web service url}
\usage{
sag_api(service, ...)
}
\arguments{
\item{service}{the name of the service}

\item{...}{name arguments will be added as queries}
}
\value{
a complete url as a character string
}
\description{
utility to build a url with optional query arguments
}
\examples{

sag_api("hi", bye = 21)
sag_api("StockList", year = 2021)

}
