% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAGTypeChartsSettings.R
\name{SAGTypeChartsandSettings}
\alias{SAGTypeChartsandSettings}
\alias{SAGTypeCharts}
\alias{SettingsForChartType}
\title{Get Details on SAG Charts and Settings}
\usage{
SAGTypeCharts(...)

SettingsForChartType(SAGChartKey, ...)
}
\arguments{
\item{...}{arguments passed to \code{\link{ices_get}}.}

\item{SAGChartKey}{the type identifier of the SAG chart, e.g. 0, 1, 2, ...}
}
\value{
a data frame with SAG chart type IDs and settings IDs.
}
\description{
List all possible chart settings for each chart type (0 = general, 1 = Landings etc.).
}
\examples{
\dontrun{
SAGTypeCharts()

SettingsForChartType(0)[-4]
}
}
