% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPackageOptions.R
\name{setPackageOptions}
\alias{setPackageOptions}
\alias{sag_use_token}
\alias{sag_messages}
\title{Get and Set icesSAG package options}
\usage{
sag_use_token(value)

sag_messages(value)
}
\arguments{
\item{value}{TRUE or FALSE}
}
\value{
invisible return of the old value.
}
\description{
There are two options of interest, 1) switch on or off the use off
authenticated web service calls, and 2) switch on or off the display of
messages to the console.
}
\examples{
\dontrun{
sag_use_token(TRUE)
sag_messages(TRUE)
}

}
