% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StockDownload.R
\name{StockDownload}
\alias{StockDownload}
\title{Get Source Data}
\usage{
StockDownload(assessmentKey, ...)
}
\arguments{
\item{assessmentKey}{the unique identifier of the stock assessment,
can be a vector}

\item{...}{arguments passed to \code{\link{ices_get}}.}
}
\value{
A data frame.
}
\description{
Get a copy of the source data for the specified stocks.
}
\examples{
\dontrun{
assessmentKey <- findAssessmentKey("had.27.46a20", year = 2022)
sourcedat <- StockDownload(assessmentKey)
head(sourcedat)
}

}
\seealso{
\code{\link{getSAG}} supports querying many years in one function call.

\code{\link{StockList}} and \code{\link{getFishStockReferencePoints}} get
  a list of stocks and reference points.

\code{\link{icesSAG-package}} gives an overview of the package.
}
\author{
Colin Millar.
}
