% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StockList.R
\name{StockList}
\alias{StockList}
\title{Get a List of Fish Stock Assessments}
\usage{
StockList(year, modifiedAfter = NULL, ...)
}
\arguments{
\item{year}{the assessment year, e.g. 2022, or 0 to process all years.}

\item{modifiedAfter}{date-time parameter in the format "". If set
will only return stocks assessments modified after the provided date.}

\item{...}{arguments passed to \code{\link{ices_get}}.}
}
\value{
A data frame.
}
\description{
Get a list of fish stock assessments for a given assessment year.
If an authentication token is not provided, only published stocks
are returned.
}
\examples{
\dontrun{
 stock_list <- StockList(year = 2022)
}
}
\seealso{
\code{\link{getSummaryTable}} gets a summary table of historical stock size.

\code{\link{getFishStockReferencePoints}} gets biological reference points.

\code{\link{icesSAG-package}} gives an overview of the package.
}
