% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StockSettings.R
\name{getsetStockChartSettings}
\alias{getsetStockChartSettings}
\alias{StockSettings}
\alias{setStockSettings}
\title{Get and Set SAG Chart Settings}
\usage{
StockSettings(assessmentKey, ...)

setStockSettings(
  assessmentKey,
  chartKey,
  settingKey,
  settingValue,
  copyNextYear,
  ...
)
}
\arguments{
\item{assessmentKey}{the unique identifier of the stock assessment}

\item{...}{arguments passed to \code{\link{ices_get}}.}

\item{chartKey}{the type identifier of the SAG chart, e.g. 0, 1, 2, ...}

\item{settingKey}{the type identifier of the SAG chart setting, e.g. 0, 1, 2, ...}

\item{settingValue}{the vale of the setting}

\item{copyNextYear}{should the settings be copied to next year (TRUE) or not (FALSE)}
}
\value{
A data frame with SAG chart type IDs, settings IDs and setting values.
}
\description{
details
}
\examples{
\dontrun{
assessmentKey <- findAssessmentKey("had.27.46a20", year = 2022)

StockSettings(assessmentKey)
}
}
