% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFishStockReferencePoints.R
\name{getFishStockReferencePoints}
\alias{getFishStockReferencePoints}
\title{Get Reference Points}
\usage{
getFishStockReferencePoints(year, stock = NULL)
}
\arguments{
\item{year}{the assessment year, e.g. 2015, or 0 to process all years.}

\item{stock}{a stock name, e.g. cod-347d, or NULL to process all stocks.}
}
\value{
A data frame.
}
\description{
Get biological reference points for all stocks in a given assessment year.
}
\examples{
refpts <- getFishStockReferencePoints(2015, "cod-347d")

}
\author{
Colin Millar and Scott Large.
}
\seealso{
\code{\link{getListStocks}} gets a list of stocks.

\code{\link{getSummaryTable}} gets a summary table of historical stock size.

\code{\link{icesSAG-package}} gives an overview of the package.
}

