% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFishStockReferencePoints.R
\name{getFishStockReferencePoints}
\alias{getFishStockReferencePoints}
\title{Query fish stock reference points for a stock}
\usage{
getFishStockReferencePoints(year)
}
\arguments{
\item{year}{the numeric year of the biological reference points output, e.g. 2010. All assessment years can be queried with 0.}
}
\value{
A data.frame.
}
\description{
Returns the biological reference points for all published stocks in a year
}
\examples{
getFishStockReferencePoints(year = 2015)


}
\author{
Colin Millar and Scott Large
}
\seealso{
\code{\link{getListStocks}} returns data frame of fish stocks for a given year 

\code{\link{getSummaryTable}} returns stock assessment summary table of all published stocks for a given year.

\code{\link{icesSAG-package}} gives an overview of the package.
}

