% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFishStockReferencePoints.R
\name{getFishStockReferencePoints}
\alias{getFishStockReferencePoints}
\title{Get Reference Points}
\usage{
getFishStockReferencePoints(key)
}
\arguments{
\item{key}{the unique identifier of the stock assessment}
}
\value{
A data frame.
}
\description{
Get biological reference points for all stocks in a given assessment year.
}
\examples{
stocklist <- getListStocks(2015)
id <- grep("cod-347d", stocklist$FishStockName)
stocklist[id,]
key <- stocklist$key[id]
refpts <- getFishStockReferencePoints(key)
refpts

}
\author{
Colin Millar and Scott Large.
}
\seealso{
\code{\link{getListStocks}} gets a list of stocks.

\code{\link{getSummaryTable}} gets a summary table of historical stock size.

\code{\link{icesSAG-package}} gives an overview of the package.
}

