% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGraph.R
\name{getStandardAssessmentGraphs}
\alias{getStandardAssessmentGraphs}
\alias{getLandingsGraph}
\alias{getRecruitmentGraph}
\alias{getFishingMortalityGraph}
\alias{getSpawningStockBiomassGraph}
\alias{getFishMortality}
\alias{getstock_recruitment}
\alias{getYSSB}
\alias{getSSBHistoricalPerformance}
\alias{getFishingMortalityHistoricalPerformance}
\alias{getRecruitmentHistoricalPerformance}
\alias{getStockStatusTable}
\alias{getCustomGraph}
\title{Get a Graph of Stock Assessment Output}
\usage{
getLandingsGraph(assessmentKey, ...)

getRecruitmentGraph(assessmentKey, ...)

getFishingMortalityGraph(assessmentKey, ...)

getSpawningStockBiomassGraph(assessmentKey, ...)

getFishMortality(assessmentKey, ...)

getstock_recruitment(assessmentKey, ...)

getYSSB(assessmentKey, ...)

getSSBHistoricalPerformance(assessmentKey, ...)

getFishingMortalityHistoricalPerformance(assessmentKey, ...)

getRecruitmentHistoricalPerformance(assessmentKey, ...)

getStockStatusTable(assessmentKey, ...)

getCustomGraph(assessmentKey, ...)
}
\arguments{
\item{assessmentKey}{the unique identifier of the stock assessment}

\item{...}{if using getCustomGraph, the custom_graph argument can be used to specify which custom graph to retrieve (see details)s).}
}
\value{
An array representing a bitmap.
}
\description{
Get a graph of stock assessment output, e.g., historical stock size,
recruitment, and fishing pressure.
}
\details{
Some stocks have custom graphs, to access these use the function
\code{getCustomGraph(assessmentKey, custom_graph)}
where custom_graph is an integer: 1, 2, 3, or 4.
}
\examples{
\dontrun{
assessmentKeys <- findAssessmentKey("had", 2015)
landings_img <- getLandingsGraph(assessmentKeys[1])
landings_img

landings_plots <- getLandingsGraph(assessmentKeys)
landings_plots
}

}
\seealso{
\code{\link{getListStocks}} gets a list of stocks.

\code{\link{getFishStockReferencePoints}} gets biological reference points.

\code{\link{icesSAG-package}} gives an overview of the package.
}
