% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLandingsGraph.R
\name{getLandingsGraph}
\alias{getLandingsGraph}
\title{Get a Summary Table of Historical Stock Size}
\usage{
getLandingsGraph(key)
}
\arguments{
\item{key}{the unique identifier of the stock assessment}
}
\value{
An array representing a png.
}
\description{
Get a summary table of historical stock size, recruitment, and fishing pressure.
}
\examples{
stocklist <- getListStocks(2016)
id <- grep("cod-2224", stocklist$FishStockName)
stocklist[id,]
key <- stocklist$key[id]
landings_img <- getLandingsGraph(key)
plot.new()
plot(landings_img)

}
\author{
Colin Millar and Scott Large.
}
\seealso{
\code{\link{getListStocks}} gets a list of stocks.

\code{\link{getFishStockReferencePoints}} gets biological reference points.

\code{\link{icesSAG-package}} gives an overview of the package.
}

