% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGraph.R
\name{getLandingsGraph}
\alias{getLandingsGraph}
\title{Get graphs of stock assessment output}
\usage{
getLandingsGraph(key)
}
\arguments{
\item{key}{the unique identifier of the stock assessment}
}
\value{
An array representing a png.
}
\description{
Get summary graphs of, e.g., historical stock size, recruitment, and fishing pressure.
}
\examples{
key <- findKey("cod-scow", 2015)
landings_img <- getLandingsGraph(key)
plot(landings_img)

keys <- findKey("had", 2015)

landings_plots <- getLandingsGraph(keys)
plot(landings_plots)

}
\seealso{
\code{\link{getListStocks}} gets a list of stocks.

\code{\link{getFishStockReferencePoints}} gets biological reference points.

\code{\link{icesSAG-package}} gives an overview of the package.
}

