% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getListStocks.R
\name{getListStocks}
\alias{getListStocks}
\title{Query list of fish stocks in the database}
\usage{
getListStocks(year)
}
\arguments{
\item{year}{the numeric year of the stock assessment output, e.g. 2010. All assessment years can be queried with 0.}
}
\value{
A data.frame.
}
\description{
Returns data frame of fish stocks for a given year (or all years when getListStocks(year = 0))
}
\examples{
\dontrun{getListStocks(year = 2015)}


}
\author{
Colin Millar and Scott Large
}
\seealso{
\code{\link{getFishStockReferencePoints}} returns the biological reference points for the stock.

\code{\link{getSummaryTable}} returns the summary table of the stock.

\code{\link{icesSAG-package}} gives an overview of the package.
}

