% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getListStocks.R
\name{getListStocks}
\alias{getListStocks}
\title{Get a List of Fish Stocks}
\usage{
getListStocks(year)
}
\arguments{
\item{year}{the assessment year, e.g. 2010. All years can be queried with 0.}
}
\value{
A data frame.
}
\description{
Get a list of fish stocks for a given assessment year.
}
\examples{
\dontrun{
stocks <- getListStocks(2015)
}

}
\author{
Colin Millar and Scott Large.
}
\seealso{
\code{\link{getFishStockReferencePoints}} gets biological reference points.

\code{\link{getSummaryTable}} gets a summary table of historical stock size.

\code{\link{icesSAG-package}} gives an overview of the package.
}

