% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getListStocks.R
\name{getListStocks}
\alias{getListStocks}
\title{Get a List of Fish Stocks}
\usage{
getListStocks(year)
}
\arguments{
\item{year}{the assessment year, e.g. 2015, or 0 to process all years.}
}
\value{
A data frame.
}
\description{
Get a list of fish stocks for a given assessment year.
}
\examples{
stocks <- getListStocks(2015)
cod.347d <- getListStocks(2015, "cod-347d")
print.simple.list(cod.347d)

}
\author{
Colin Millar, Scott Large, and Arni Magnusson.
}
\seealso{
\code{\link{getSummaryTable}} gets a summary table of historical stock size.

\code{\link{getFishStockReferencePoints}} gets biological reference points.

\code{\link{icesSAG-package}} gives an overview of the package.
}

