% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findGraphSettings.R,
%   R/getSAGTypeGraphsSettings.R
\name{getSAGGTypegraphsandSettings}
\alias{getSAGGTypegraphsandSettings}
\alias{getSAGTypeGraphs}
\alias{getSAGTypeSettings}
\title{Find Details on SAG Charts and Settings}
\usage{
getSAGTypeGraphs()

getSAGTypeSettings(SAGChartKey)
}
\arguments{
\item{SAGChartKey}{the type identifier of the SAG chart, e.g. 0, 1, 2, ...}
}
\value{
a data frame with SAG chart type IDs and settings IDs.

a data frame with SAG chart type IDs and settings IDs.
}
\description{
Provide the settings for a SAG graph given a textual name of
the graph.

List all possible chart settings for each chart type (0 = general, 1 = Landings etc.).
}
\examples{
\dontrun{
getSAGTypeGraphs()

getSAGTypeSettings(0)[-4]

key <- findAssessmentKey("cod.27.47d20", 2021)
getSAGSettingsForAStock(key[1])
}
\dontrun{
getSAGTypeGraphs()

getSAGTypeSettings(0)[-4]
}
}
