% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummaryTable.R
\name{getSummaryTable}
\alias{getSummaryTable}
\title{Get a Summary Table of Historical Stock Size}
\usage{
getSummaryTable(year)
}
\arguments{
\item{year}{the assessment year, e.g. 2010. All years can be queried with 0.}
}
\value{
A data frame.
}
\description{
Get a summary table of historical stock size, recruitment, and fishing pressure.
}
\examples{
\dontrun{
sumtab <- getSummaryTable(2015)
}

}
\author{
Colin Millar and Scott Large.
}
\seealso{
\code{\link{getListStocks}} gets a list of stocks.

\code{\link{getFishStockReferencePoints}} gets biological reference points.

\code{\link{icesSAG-package}} gives an overview of the package.
}

