% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummaryTable.R
\name{getSummaryTable}
\alias{getSummaryTable}
\title{Get a Summary Table of Historical Stock Size}
\usage{
getSummaryTable(year, stock = NULL)
}
\arguments{
\item{year}{the assessment year, e.g. 2015, or 0 to process all years.}

\item{stock}{a stock name, e.g. cod-347d, or NULL to process all stocks.}
}
\value{
A data frame.
}
\description{
Get a summary table of historical stock size, recruitment, and fishing pressure.
}
\examples{
sumtab <- getSummaryTable(2015, "cod-347d")

}
\author{
Colin Millar and Scott Large.
}
\seealso{
\code{\link{getListStocks}} gets a list of stocks.

\code{\link{getFishStockReferencePoints}} gets biological reference points.

\code{\link{icesSAG-package}} gives an overview of the package.
}

