% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesSAG-package.R
\docType{package}
\name{icesSAG-package}
\alias{icesSAG-package}
\alias{icesSAG}
\title{Stock Assessment Graphs Database Web Services}
\description{
R interface to access the web services of the ICES (International Council
for the Exploration of the Sea) Stock Assessment Graphs (SAG) database.
}
\details{
\emph{Get dataset:}
\tabular{ll}{
  \code{\link{getSummaryTable}}             \tab summary results\cr
  \code{\link{getFishStockReferencePoints}} \tab reference points\cr
  \code{\link{getSAG}}                      \tab any data
}
\emph{Look up codes:}
\tabular{ll}{
  \code{\link{findKey}}       \tab find key\cr
  \code{\link{getListStocks}} \tab list of stocks
}
}
\references{
ICES Stock Assessment Graphs database: \url{http://sg.ices.dk}.
}
\author{
Colin Millar, Scott Large, and Arni Magnusson.
}
