% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesSAG-package.R
\docType{package}
\name{icesSAG-package}
\alias{icesSAG}
\alias{icesSAG-package}
\title{Stock Assessment Graphs Database Web Services}
\description{
R interface to access the web services of the ICES (International Council
for the Exploration of the Sea) Stock Assessment Graphs (SAG) database.
}
\details{
\emph{Get dataset:}
\tabular{ll}{
  \code{\link{getFishStockReferencePoints}} \tab reference points\cr
  \code{\link{getListStocks}}               \tab list of stocks\cr
  \code{\link{getSummaryTable}}             \tab historical stock size
}
}
\author{
Colin Millar, Scott Large, and Arni Magnusson.
}
\references{
ICES Stock Assessment Graphs:
\url{http://ices.dk/marine-data/tools/Pages/stock-assessment-graphs.aspx}.

ICES software applications:
\url{http://ices.dk/marine-data/tools/Pages/Software.aspx}.
}

