% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesSAG-package.R
\docType{package}
\name{icesSAG-package}
\alias{icesSAG-package}
\alias{icesSAG}
\title{Stock Assessment Graphs Database Web Services}
\description{
R interface to access the web services of the ICES Stock Assessment Graphs
database.
}
\details{
\emph{Get dataset:}
\tabular{ll}{
  \code{\link{getSummaryTable}}             \tab summary results\cr
  \code{\link{getFishStockReferencePoints}} \tab reference points\cr
  \code{\link{getSAG}}                      \tab any data
}
\emph{Look up codes:}
\tabular{ll}{
  \code{\link{findAssessmentKey}}       \tab find assessment key\cr
  \code{\link{getListStocks}} \tab list of stocks
}
}
\references{
ICES Stock Assessment Graphs database: \url{http://sg.ices.dk}.

ICES Stock Assessment Graphs web services:
\url{http://sg.ices.dk/webservices.aspx}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sg.ices.dk}
  \item \url{https://github.com/ices-tools-prod/icesSAG}
  \item Report bugs at \url{https://github.com/ices-tools-prod/icesSAG/issues}
}

}
\author{
Colin Millar, Scott Large, Arni Magnusson, Carlos Pinto and Laura Andreea Petre.
}
