% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadStock.R
\name{uploadStock}
\alias{uploadStock}
\title{Upload New or Updated Fish Stock Assessment Results}
\usage{
uploadStock(info, fishdata, verbose = FALSE)
}
\arguments{
\item{info}{a list of stock information}

\item{fishdata}{a data.frame of fish data}

\item{verbose}{if TRUE more verbose messages are reported}
}
\value{
The database key of the new / updated stock, or 0 if there was an error.
}
\description{
Get summary results of historical stock size, recruitment, and fishing
pressure.
}
\seealso{
\code{\link{stockInfo}} creates a list of stock information.

\code{\link{stockFishdata}} creates a data.frame of fish stock summary data.
}
\author{
Colin Millar.
}
